servo_media_audio::audio_node

Trait AudioNodeEngine

pub(crate) trait AudioNodeEngine: Send + AudioNodeCommon {

    // Required methods
    fn node_type(&self) -> AudioNodeType;
    fn process(&mut self, inputs: Chunk, info: &BlockInfo) -> Chunk;

    // Provided methods
    fn mute_node(&mut self) -> Chunk { ... }
    fn message(&mut self, msg: AudioNodeMessage, sample_rate: f32) { ... }
    fn message_specific(&mut self, _: AudioNodeMessage, _sample_rate: f32) { ... }
    fn input_count(&self) -> u32 { ... }
    fn output_count(&self) -> u32 { ... }
    fn channel_count(&self) -> u8 { ... }
    fn channel_count_mode(&self) -> ChannelCountMode { ... }
    fn channel_interpretation(&self) -> ChannelInterpretation { ... }
    fn set_channel_interpretation(&mut self, i: ChannelInterpretation) { ... }
    fn set_channel_count(&mut self, c: u8) { ... }
    fn set_channel_count_mode(&mut self, m: ChannelCountMode) { ... }
    fn destination_data(&mut self) -> Option<Chunk> { ... }
    fn get_param(&mut self, _: ParamType) -> &mut Param { ... }
    fn set_listenerdata(&mut self, _: Block) { ... }

}

This trait represents the common features of all audio nodes.

Required Methods

fn node_type(&self) -> AudioNodeType

fn process(&mut self, inputs: Chunk, info: &BlockInfo) -> Chunk

Provided Methods

fn mute_node(&mut self) -> Chunk

https://webaudio.github.io/web-audio-api/#mute

Muting an AudioNode means that its output MUST be silence for the rendering of this audio block.

fn message(&mut self, msg: AudioNodeMessage, sample_rate: f32)

fn message_specific(&mut self, _: AudioNodeMessage, _sample_rate: f32)

Messages specific to this node

fn input_count(&self) -> u32

fn output_count(&self) -> u32

fn channel_count(&self) -> u8

Number of input channels for each input port

fn channel_count_mode(&self) -> ChannelCountMode

fn channel_interpretation(&self) -> ChannelInterpretation

fn set_channel_interpretation(&mut self, i: ChannelInterpretation)

fn set_channel_count(&mut self, c: u8)

fn set_channel_count_mode(&mut self, m: ChannelCountMode)

fn destination_data(&mut self) -> Option<Chunk>

If we’re the destination node, extract the contained data

fn get_param(&mut self, _: ParamType) -> &mut Param

fn set_listenerdata(&mut self, _: Block)