Module servo_media_audio::graph

Structs

AudioGraph

Connection
A single connection between ports

Edge
An edge in the graph

InputPort
Marker type for denoting that the port is an input port of the node it is connected to

Node

NodeId
A unique identifier for nodes in the graph. Stable under graph mutation.

OutputPort
Marker type for denoting that the port is an output port of the node it is connected to

PortId
An identifier for a port.

Enums

PortIndex
A zero-indexed "port" for a node. Most nodes have one input and one output port, but some may have more. For example, a channel splitter node will have one output port for each channel.

Void

Traits

PortKind