Enum servo_media_audio::graph::PortIndex

pub enum PortIndex<Kind: PortKind> {
    Port(u32),
    Param(Kind::ParamId),
    Listener(Kind::Listener),
}

A zero-indexed “port” for a node. Most nodes have one input and one output port, but some may have more. For example, a channel splitter node will have one output port for each channel.

These are essentially indices into the Chunks

Kind is a zero sized type and is useful for distinguishing between input and output ports (which may otherwise share indices)

Variants

Port(u32)

Param(Kind::ParamId)

Listener(Kind::Listener)

special variant only used for the implicit connection from listeners to params