Struct servo_media_audio::graph::AudioGraph

pub struct AudioGraph {
    graph: StableGraph<Node, Edge>,
    dest_id: NodeId,
    dests: Vec<NodeId>,
    listener_id: NodeId,
}

Fields
graph: StableGraph<Node, Edge>dest_id: NodeIddests: Vec<NodeId>listener_id: NodeId
Implementations

impl AudioGraph

pub fn new(channel_count: u8) -> Self

pub(crate) fn add_node(&mut self, node: Box<dyn AudioNodeEngine>) -> NodeId

Create a node, obtain its id

pub fn add_edge(&mut self, out: PortId<OutputPort>, inp: PortId<InputPort>)

Connect an output port to an input port

The edge goes from the output port to the input port, connecting two nodes

pub fn disconnect_all_from(&mut self, node: NodeId)

Disconnect all outgoing connections from a node

https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect

pub fn disconnect_output(&mut self, out: PortId<OutputPort>)

pub fn disconnect_between(&mut self, from: NodeId, to: NodeId)

Disconnect connections from a node to another node

https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect-destinationnode

pub fn disconnect_output_between(&mut self, out: PortId<OutputPort>, to: NodeId)

Disconnect all outgoing connections from a node’s output to another node

https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect-destinationnode-output

pub fn disconnect_to(&mut self, node: NodeId, inp: PortId<InputPort>)

Disconnect all outgoing connections from a node to another node’s input

Only used in WebAudio for disconnecting audio params

https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect-destinationparam

pub fn disconnect_output_between_to(
&mut self,
out: PortId<OutputPort>,
inp: PortId<InputPort>
)

Disconnect all outgoing connections from a node’s output to another node’s input

https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect-destinationnode-output-input https://webaudio.github.io/web-audio-api/#dom-audionode-disconnect-destinationparam-output

pub fn dest_id(&self) -> NodeId

Get the id of the destination node in this graph

All graphs have a destination node, with one input port

pub fn add_extra_dest(&mut self, dest: NodeId)

Add additional terminator nodes

pub fn listener_id(&self) -> NodeId

Get the id of the AudioListener in this graph

All graphs have a single listener, with no ports (but nine AudioParams)

N.B. The listener actually has a single output port containing its position data for the block, however this should not be exposed to the DOM.

pub fn process(&mut self, info: &BlockInfo) -> Chunk

For a given block, process all the data on this graph

pub(crate) fn node_mut(
&self,
ix: NodeId
) -> RefMut<'_, Box<dyn AudioNodeEngine>>

Obtain a mutable reference to a node