Struct servo_media_audio::param::Param

pub struct Param {
    val: f32,
    kind: ParamRate,
    events: Vec<AutomationEvent>,
    current_event: usize,
    event_start_time: Tick,
    event_start_value: f32,
    blocks: Vec<Block>,
    block_mix_val: f32,
    summed: bool,
    dirty: bool,
}

An AudioParam.

https://webaudio.github.io/web-audio-api/#AudioParam

Fields
val: f32kind: ParamRateevents: Vec<AutomationEvent>current_event: usizeevent_start_time: Tickevent_start_value: f32blocks: Vec<Block>

Cache of inputs from connect()ed nodes

block_mix_val: f32

The value of all connect()ed inputs mixed together, for this frame

summed: bool

If true, blocks has been summed together into a single block

dirty: bool
Implementations

impl Param

pub fn new(val: f32) -> Self

pub fn new_krate(val: f32) -> Self

pub fn update(&mut self, block: &BlockInfo, tick: Tick) -> bool

Update the value of this param to the next

Invariant: This should be called with monotonically increasing ticks, and Tick(0) should never be skipped.

Returns true if anything changed

pub fn value(&self) -> f32

pub fn set_rate(&mut self, rate: ParamRate)

pub(crate) fn insert_event(&mut self, event: AutomationEvent)

pub(crate) fn add_block(&mut self, block: Block)

pub(crate) fn flush_to_block(&mut self, info: &BlockInfo, block: &mut [f32])

Flush an entire block of values into a buffer

Only for use with AudioListener.

Invariant: block must be a FRAMES_PER_BLOCK length array filled with silence